+++ .gitignore
@@ -24,7 +24,7 @@ migrate_working_dir/
 # Flutter/Dart/Pub related
 **/doc/api/
 **/ios/Flutter/.last_build_id
-../.dart_tool/
+.dart_tool/
 .flutter-plugins
 .flutter-plugins-dependencies
 .pub-cache/
