--- lib/app.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class App extends StatelessWidget {
-  const App({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Placeholder(),
-    );
-  }
-}
