+++ lib/config/app.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 
+import '../presentation/main_screen.dart';
+
 class App extends StatelessWidget {
   const App({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(home: Placeholder());
+    return const MaterialApp(home: MainScreen());
   }
 }
