+++ lib/presentation/main_screen.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+import '../main.dart';
+
+class MainScreen extends StatelessWidget {
+  const MainScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MyHomePage(title: "",);
+  }
+}
